+++ Scala-Play-OneStopWeb/app/models/CustomerOrderItem.scala
@@ -2,7 +2,7 @@ package models
 
 import anorm._
 
-case class CustomerOrderItem( customerOrderId: Pk[Long], itemId: Long, quantity: Int, unitPrice: BigDecimal )
+case class CustomerOrderItem( customerOrderId: Pk[Long], itemId: Long, quantity: Long, unitPrice: BigDecimal )
 
 object CustomerOrderItem {
